+++ docs/merkletree/treetex/main.go
@@ -379,7 +379,8 @@ func main() {
 	if *inclusion > 0 {
 		leafID := compact.NewNodeID(0, uint64(*inclusion))
 		modifyNodeInfo(leafID, func(n *nodeInfo) { n.incPath = true })
-		nf, err := merkle.CalcInclusionProofNodeAddresses(int64(*treeSize), *inclusion, int64(*treeSize))
+		// TODO(pavelkalinnikov): Highlight the "ephemeral" node too.
+		nf, err := merkle.CalcInclusionProofNodeAddresses(int64(*treeSize), *inclusion)
 		if err != nil {
 			log.Fatalf("Failed to calculate inclusion proof addresses: %s", err)
 		}