+++ server/log_rpc_server_test.go
@@ -107,10 +107,11 @@ var (
 	nodeIdsInclusionSize7Index2 = []compact.NodeID{
 		compact.NewNodeID(0, 3),
 		compact.NewNodeID(1, 0),
-		compact.NewNodeID(2, 1),
+		compact.NewNodeID(0, 6),
+		compact.NewNodeID(1, 2),
 	}
 
-	nodeIdsConsistencySize4ToSize7 = []compact.NodeID{compact.NewNodeID(2, 1)}
+	nodeIdsConsistencySize4ToSize7 = []compact.NodeID{compact.NewNodeID(0, 6), compact.NewNodeID(1, 2)}
 	corruptLogRoot                 = &trillian.SignedLogRoot{LogRoot: []byte("this is not tls encoded data")}
 )
 
@@ -598,7 +599,7 @@ func TestGetProofByHashErrors(t *testing.T) {
 				tx.EXPECT().Close().Return(nil)
 			},
 			req:    &getInclusionProofByHashRequest7,
-			errStr: "expected 3 nodes",
+			errStr: "expected 4 nodes",
 		},
 		{
 			name: "wrong node",
@@ -608,7 +609,10 @@ func TestGetProofByHashErrors(t *testing.T) {
 				tx.EXPECT().LatestSignedLogRoot(gomock.Any()).Return(signedRoot1, nil)
 				tx.EXPECT().GetLeavesByHash(gomock.Any(), [][]byte{leafHash1}, false).Return([]*trillian.LogLeaf{{LeafIndex: 2}}, nil)
 				// We set this up so one of the returned nodes has the wrong ID
-				tx.EXPECT().GetMerkleNodes(gomock.Any(), nodeIdsInclusionSize7Index2).Return([]tree.Node{{ID: nodeIdsInclusionSize7Index2[0]}, {ID: compact.NewNodeID(4, 5)}, {ID: nodeIdsInclusionSize7Index2[2]}}, nil)
+				tx.EXPECT().GetMerkleNodes(gomock.Any(), nodeIdsInclusionSize7Index2).Return([]tree.Node{
+					{ID: nodeIdsInclusionSize7Index2[0]}, {ID: compact.NewNodeID(4, 5)},
+					{ID: nodeIdsInclusionSize7Index2[2]}, {ID: nodeIdsInclusionSize7Index2[3]},
+				}, nil)
 				tx.EXPECT().Close().Return(nil)
 			},
 			req:    &getInclusionProofByHashRequest7,
@@ -712,6 +716,7 @@ func TestGetProofByHash(t *testing.T) {
 				{ID: nodeIdsInclusionSize7Index2[0], Hash: []byte("nodehash0")},
 				{ID: nodeIdsInclusionSize7Index2[1], Hash: []byte("nodehash1")},
 				{ID: nodeIdsInclusionSize7Index2[2], Hash: []byte("nodehash2")},
+				{ID: nodeIdsInclusionSize7Index2[3], Hash: []byte("nodehash3")},
 			}, nil).AnyTimes()
 			mockTX.EXPECT().Commit(gomock.Any()).Return(nil)
 			mockTX.EXPECT().Close().Return(nil)
@@ -747,7 +752,7 @@ func TestGetProofByHash(t *testing.T) {
 				Hashes: [][]byte{
 					[]byte("nodehash0"),
 					[]byte("nodehash1"),
-					[]byte("nodehash2"),
+					th.HashChildren([]byte("nodehash3"), []byte("nodehash2")),
 				},
 			}
 
@@ -825,7 +830,7 @@ func TestGetProofByIndex(t *testing.T) {
 				tx.EXPECT().Close().Return(nil)
 			},
 			req:    &getInclusionProofByIndexRequest7,
-			errStr: "expected 3 nodes",
+			errStr: "expected 4 nodes",
 		},
 		{
 			name: "wrong node",
@@ -834,7 +839,10 @@ func TestGetProofByIndex(t *testing.T) {
 				s.EXPECT().SnapshotForTree(gomock.Any(), cmpMatcher{tree1}).Return(tx, nil)
 				// We set this up so one of the returned nodes has the wrong ID
 				tx.EXPECT().LatestSignedLogRoot(gomock.Any()).Return(signedRoot1, nil)
-				tx.EXPECT().GetMerkleNodes(gomock.Any(), nodeIdsInclusionSize7Index2).Return([]tree.Node{{ID: nodeIdsInclusionSize7Index2[0]}, {ID: compact.NewNodeID(4, 5)}, {ID: nodeIdsInclusionSize7Index2[2]}}, nil)
+				tx.EXPECT().GetMerkleNodes(gomock.Any(), nodeIdsInclusionSize7Index2).Return([]tree.Node{
+					{ID: nodeIdsInclusionSize7Index2[0]}, {ID: compact.NewNodeID(4, 5)},
+					{ID: nodeIdsInclusionSize7Index2[2]}, {ID: nodeIdsInclusionSize7Index2[3]},
+				}, nil)
 				tx.EXPECT().Close().Return(nil)
 			},
 			req:    &getInclusionProofByIndexRequest7,
@@ -846,7 +854,10 @@ func TestGetProofByIndex(t *testing.T) {
 				tx := storage.NewMockLogTreeTX(c)
 				s.EXPECT().SnapshotForTree(gomock.Any(), cmpMatcher{tree1}).Return(tx, nil)
 				tx.EXPECT().LatestSignedLogRoot(gomock.Any()).Return(signedRoot1, nil)
-				tx.EXPECT().GetMerkleNodes(gomock.Any(), nodeIdsInclusionSize7Index2).Return([]tree.Node{{ID: nodeIdsInclusionSize7Index2[0]}, {ID: nodeIdsInclusionSize7Index2[1]}, {ID: nodeIdsInclusionSize7Index2[2]}}, nil)
+				tx.EXPECT().GetMerkleNodes(gomock.Any(), nodeIdsInclusionSize7Index2).Return([]tree.Node{
+					{ID: nodeIdsInclusionSize7Index2[0]}, {ID: nodeIdsInclusionSize7Index2[1]},
+					{ID: nodeIdsInclusionSize7Index2[2]}, {ID: nodeIdsInclusionSize7Index2[3]},
+				}, nil)
 				tx.EXPECT().Commit(gomock.Any()).Return(errors.New("COMMIT"))
 				tx.EXPECT().Close().Return(nil)
 			},
@@ -885,6 +896,7 @@ func TestGetProofByIndex(t *testing.T) {
 					{ID: nodeIdsInclusionSize7Index2[0], Hash: []byte("nodehash0")},
 					{ID: nodeIdsInclusionSize7Index2[1], Hash: []byte("nodehash1")},
 					{ID: nodeIdsInclusionSize7Index2[2], Hash: []byte("nodehash2")},
+					{ID: nodeIdsInclusionSize7Index2[3], Hash: []byte("nodehash3")},
 				}, nil)
 				tx.EXPECT().Commit(gomock.Any()).Return(nil)
 				tx.EXPECT().Close().Return(nil)
@@ -897,7 +909,7 @@ func TestGetProofByIndex(t *testing.T) {
 					Hashes: [][]byte{
 						[]byte("nodehash0"),
 						[]byte("nodehash1"),
-						[]byte("nodehash2"),
+						th.HashChildren([]byte("nodehash3"), []byte("nodehash2")),
 					},
 				},
 			},
@@ -995,6 +1007,7 @@ func TestGetEntryAndProof(t *testing.T) {
 					{ID: nodeIdsInclusionSize7Index2[0], Hash: []byte("nodehash0")},
 					{ID: nodeIdsInclusionSize7Index2[1], Hash: []byte("nodehash1")},
 					{ID: nodeIdsInclusionSize7Index2[2], Hash: []byte("nodehash2")},
+					{ID: nodeIdsInclusionSize7Index2[3], Hash: []byte("nodehash3")},
 				}, nil)
 				tx.EXPECT().GetLeavesByRange(gomock.Any(), int64(2), int64(1)).Return(nil, errors.New("STORAGE"))
 				tx.EXPECT().Close().Return(nil)
@@ -1012,6 +1025,7 @@ func TestGetEntryAndProof(t *testing.T) {
 					{ID: nodeIdsInclusionSize7Index2[0], Hash: []byte("nodehash0")},
 					{ID: nodeIdsInclusionSize7Index2[1], Hash: []byte("nodehash1")},
 					{ID: nodeIdsInclusionSize7Index2[2], Hash: []byte("nodehash2")},
+					{ID: nodeIdsInclusionSize7Index2[3], Hash: []byte("nodehash3")},
 				}, nil)
 				tx.EXPECT().GetLeavesByRange(gomock.Any(), int64(2), int64(1)).Return([]*trillian.LogLeaf{leaf1}, nil)
 				tx.EXPECT().Commit(gomock.Any()).Return(errors.New("COMMIT"))
@@ -1052,6 +1066,7 @@ func TestGetEntryAndProof(t *testing.T) {
 					{ID: nodeIdsInclusionSize7Index2[0], Hash: []byte("nodehash0")},
 					{ID: nodeIdsInclusionSize7Index2[1], Hash: []byte("nodehash1")},
 					{ID: nodeIdsInclusionSize7Index2[2], Hash: []byte("nodehash2")},
+					{ID: nodeIdsInclusionSize7Index2[3], Hash: []byte("nodehash3")},
 				}, nil)
 				// Code passed one leaf index so expects one result, but we return more
 				tx.EXPECT().GetLeavesByRange(gomock.Any(), int64(2), int64(1)).Return([]*trillian.LogLeaf{leaf1, leaf3}, nil)
@@ -1070,6 +1085,7 @@ func TestGetEntryAndProof(t *testing.T) {
 					{ID: nodeIdsInclusionSize7Index2[0], Hash: []byte("nodehash0")},
 					{ID: nodeIdsInclusionSize7Index2[1], Hash: []byte("nodehash1")},
 					{ID: nodeIdsInclusionSize7Index2[2], Hash: []byte("nodehash2")},
+					{ID: nodeIdsInclusionSize7Index2[3], Hash: []byte("nodehash3")},
 				}, nil)
 				tx.EXPECT().GetLeavesByRange(gomock.Any(), int64(2), int64(1)).Return([]*trillian.LogLeaf{leaf1}, nil)
 				tx.EXPECT().Commit(gomock.Any()).Return(nil)
@@ -1083,7 +1099,7 @@ func TestGetEntryAndProof(t *testing.T) {
 					Hashes: [][]byte{
 						[]byte("nodehash0"),
 						[]byte("nodehash1"),
-						[]byte("nodehash2"),
+						th.HashChildren([]byte("nodehash3"), []byte("nodehash2")),
 					},
 				},
 				Leaf: leaf1,
@@ -1113,6 +1129,7 @@ func TestGetEntryAndProof(t *testing.T) {
 					{ID: nodeIdsInclusionSize7Index2[0], Hash: []byte("nodehash0")},
 					{ID: nodeIdsInclusionSize7Index2[1], Hash: []byte("nodehash1")},
 					{ID: nodeIdsInclusionSize7Index2[2], Hash: []byte("nodehash2")},
+					{ID: nodeIdsInclusionSize7Index2[3], Hash: []byte("nodehash3")},
 				}, nil)
 				tx.EXPECT().GetLeavesByRange(gomock.Any(), int64(2), int64(1)).Return([]*trillian.LogLeaf{leaf1}, nil)
 				tx.EXPECT().Commit(gomock.Any()).Return(nil)
@@ -1126,7 +1143,7 @@ func TestGetEntryAndProof(t *testing.T) {
 					Hashes: [][]byte{
 						[]byte("nodehash0"),
 						[]byte("nodehash1"),
-						[]byte("nodehash2"),
+						th.HashChildren([]byte("nodehash3"), []byte("nodehash2")),
 					},
 				},
 				Leaf: leaf1,
@@ -1215,26 +1232,36 @@ func TestGetConsistencyProof(t *testing.T) {
 			errStr:     "commit",
 			wantHashes: [][]byte{[]byte("nodehash")},
 			nodeIDs:    nodeIdsConsistencySize4ToSize7,
-			nodes:      []tree.Node{{ID: compact.NewNodeID(2, 1), Hash: []byte("nodehash")}},
-			commitErr:  errors.New("commit() failed"),
+			nodes: []tree.Node{
+				{ID: compact.NewNodeID(0, 6), Hash: []byte("nodehash1")},
+				{ID: compact.NewNodeID(1, 2), Hash: []byte("nodehash2")},
+			},
+			commitErr: errors.New("commit() failed"),
 		},
 		{
 			// Storage doesn't return the requested node, should result in an error.
 			req:        &getConsistencyProofRequest7,
-			errStr:     "expected node {2 1} at",
+			errStr:     "expected node {0 6} at",
 			wantHashes: [][]byte{[]byte("nodehash")},
 			nodeIDs:    nodeIdsConsistencySize4ToSize7,
-			nodes:      []tree.Node{{ID: compact.NewNodeID(3, 1), Hash: []byte("nodehash")}},
-			noCommit:   true,
+			nodes: []tree.Node{
+				{ID: compact.NewNodeID(3, 1), Hash: []byte("nodehash1")},
+				{ID: compact.NewNodeID(1, 2), Hash: []byte("nodehash2")},
+			},
+			noCommit: true,
 		},
 		{
 			// Storage returns an unexpected extra node, should result in an error.
 			req:        &getConsistencyProofRequest7,
-			errStr:     "expected 1 nodes",
+			errStr:     "expected 2 nodes",
 			wantHashes: [][]byte{[]byte("nodehash")},
 			nodeIDs:    nodeIdsConsistencySize4ToSize7,
-			nodes:      []tree.Node{{ID: compact.NewNodeID(2, 1), Hash: []byte("nodehash")}, {ID: compact.NewNodeID(3, 10), Hash: []byte("nodehash2")}},
-			noCommit:   true,
+			nodes: []tree.Node{
+				{ID: compact.NewNodeID(0, 6), Hash: []byte("nodehash1")},
+				{ID: compact.NewNodeID(1, 2), Hash: []byte("nodehash2")},
+				{ID: compact.NewNodeID(3, 10), Hash: []byte("nodehash3")},
+			},
+			noCommit: true,
 		},
 		{
 			// Ask for a proof from size 4 to 8 but the tree is only size 7. This should succeed but with no proof.
@@ -1246,9 +1273,12 @@ func TestGetConsistencyProof(t *testing.T) {
 		{
 			// A normal request which should succeed.
 			req:        &getConsistencyProofRequest7,
-			wantHashes: [][]byte{[]byte("nodehash")},
+			wantHashes: [][]byte{th.HashChildren([]byte("nodehash2"), []byte("nodehash1"))},
 			nodeIDs:    nodeIdsConsistencySize4ToSize7,
-			nodes:      []tree.Node{{ID: compact.NewNodeID(2, 1), Hash: []byte("nodehash")}},
+			nodes: []tree.Node{
+				{ID: compact.NewNodeID(0, 6), Hash: []byte("nodehash1")},
+				{ID: compact.NewNodeID(1, 2), Hash: []byte("nodehash2")},
+			},
 		},
 		{
 			// Tests first==second edge case, which should succeed but is an empty proof.